Refactor QuestionContext inside App - WIP due to error Uncaught TypeError: create is not a function
